# proxy/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple golang application that proxies calls to album-service",
        "title": "Proxy Service API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9070",
    "basePath": "/",
    "paths": {
        "/albums": {
            "get": {
                "description": "get all the albums in the store",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get all Albums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Album"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "add a new album to the store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create album",
                "parameters": [
                    {
                        "description": "album",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Album"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Album"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        },
        "/albums/{id}": {
            "get": {
                "description": "get as single album by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get Album by id",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Album"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Album": {
            "type": "object",
            "required": [
                "artist",
                "price",
                "title"
            ],
            "properties": {
                "artist": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 2
                },
                "id": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "price": {
                    "type": "number",
                    "maximum": 10000,
                    "minimum": 0
                },
                "title": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 2
                }
            }
        },
        "model.BindingErrorMsg": {
            "type": "object",
            "required": [
                "field",
                "message"
            ],
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ServerError": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BindingErrorMsg"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}